Update Index building and logging configuration - Restrict Index building to only the command which requires index - Added logging level for --quiet in main.py
